+++ app/v1/modules/kubernetes_services/__init__.py
@@ -6,6 +6,7 @@ from .script_to_dict import ScriptToDictParserService
 from .script_read_volumes import ScriptVolumesReaderService
 from .script_read_env import ScriptEnvironReaderService
 from .script_read_env_from import ScriptEnvironReaderFromService
+from .script_read_image_pull_secrets import ScriptImagePullSecretsReaderService
 from .script_read_ingress_rules import ScriptIngressRulesReaderService
 from .script_read_ingress_tls import ScriptIngressTlsReaderService
 from .script_read_volumes_mounts import ScriptVolumeMountsReaderService
@@ -17,6 +18,7 @@ __all__ = [
     "ScriptVolumesReaderService",
     "ScriptEnvironReaderFromService",
     "ScriptEnvironReaderService",
+    "ScriptImagePullSecretsReaderService",
     "ScriptIngressRulesReaderService",
     "ScriptIngressTlsReaderService",
     "ScriptVolumeMountsReaderService",
